Cleanup surrounding compiler warnings.

+++ src/ParsingConsole.cpp
@@ -176,7 +176,7 @@ void ParsingConsole::fetchLog(StringBuilder* l) {
 int8_t ParsingConsole::defineCommand(const char* c, const TCode* f, const char* h, const char* p, const uint8_t r, const consoleCallback ccb) {
   ConsoleCommand* cmd = new ConsoleCommand(c, '\0', f, h, p, r, ccb);
   if (nullptr != cmd) {
-    _max_cmd_len = strict_max(_max_cmd_len, strlen(c));
+    _max_cmd_len = strict_max(_max_cmd_len, (uint8_t) strlen(c));
     _cmd_list.insert(cmd);
     return 0;
   }
@@ -187,7 +187,7 @@ int8_t ParsingConsole::defineCommand(const char* c, const TCode* f, const char* 
 int8_t ParsingConsole::defineCommand(const char* c, const char sc, const TCode* f, const char* h, const char* p, const uint8_t r, const consoleCallback ccb) {
   ConsoleCommand* cmd = new ConsoleCommand(c, sc, f, h, p, r, ccb);
   if (nullptr != cmd) {
-    _max_cmd_len = strict_max(_max_cmd_len, strlen(c));
+    _max_cmd_len = strict_max(_max_cmd_len, (uint8_t) strlen(c));
     _cmd_list.insert(cmd);
     return 0;
   }
@@ -197,7 +197,7 @@ int8_t ParsingConsole::defineCommand(const char* c, const char sc, const TCode* 
 
 int8_t ParsingConsole::defineCommand(const ConsoleCommand* cmd) {
   if (nullptr != cmd) {
-    _max_cmd_len = strict_max(_max_cmd_len, strlen(cmd->cmd));
+    _max_cmd_len = strict_max(_max_cmd_len, (uint8_t) strlen(cmd->cmd));
     _cmd_list.insert((ConsoleCommand*) cmd);
     return 0;
   }
@@ -213,7 +213,7 @@ int8_t ParsingConsole::defineCommands(const ConsoleCommand* cmds, const int cmd_
   if (nullptr != cmds) {
     for (int i = 0; i < cmd_count; i++) {
       _cmd_list.insert((ConsoleCommand*) cmds + i);
-      _max_cmd_len = strict_max(_max_cmd_len, strlen(((ConsoleCommand*) cmds + i)->cmd));
+      _max_cmd_len = strict_max(_max_cmd_len, (uint8_t) strlen(((ConsoleCommand*) cmds + i)->cmd));
     }
     return 0;
   }
